rename test files & better test codes
