+++ logger/logger_inner_test.go
@@ -50,13 +50,13 @@ func TestSetLevel(t *testing.T) {
 
 	SetLevel(DebugLevel)
 	output1 := captureOutput(func() {
-		Debugf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Debugf("hello=%s number=%d", "world", 42)
 	})
-	assert.Regexp(t, `time="[^"]+" level=debug msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output1)
+	assert.Regexp(t, `time="[^"]+" level=debug msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output1)
 
 	SetLevel(InfoLevel)
 	output2 := captureOutput(func() {
-		Debugf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Debugf("hello=%s number=%d", "world", 42)
 	})
 	assert.Equal(t, "", output2)
 }
@@ -64,16 +64,16 @@ func TestSetLevel(t *testing.T) {
 func TestSetFullpath(t *testing.T) {
 	SetFullpath(true)
 	output := captureOutput(func() {
-		Warnf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Warnf("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world lorem=ipsum number=42" file="[^"]+/common/logger/logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world number=42" file="[^"]+/common/logger/logger_inner_test.go:[0-9]+"`, output)
 	SetFullpath(false)
 	output = captureOutput(func() {
-		Warnf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Warnf("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output)
 }
 
 func captureOutput(f func()) string {
@@ -86,40 +86,40 @@ func captureOutput(f func()) string {
 
 func TestDebugf(t *testing.T) {
 	output := captureOutput(func() {
-		Debugf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Debugf("hello=%s number=%d", "world", 42)
 	})
 	assert.Equal(t, "", output)
 }
 
 func TestInfof(t *testing.T) {
 	output := captureOutput(func() {
-		Infof("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Infof("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=info msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=info msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output)
 }
 
 func TestWarnf(t *testing.T) {
 	output := captureOutput(func() {
-		Warnf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Warnf("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=warning msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output)
 }
 
 func TestErrorf(t *testing.T) {
 	output := captureOutput(func() {
-		Errorf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Errorf("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=error msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=error msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output)
 }
 
 func TestFatalf(t *testing.T) {
 	_, output, err := tester.CaptureChildTest(func() {
-		Fatalf("hello=%s lorem=%s number=%d", "world", "ipsum", 42)
+		Fatalf("hello=%s number=%d", "world", 42)
 	})
 	t.Log(output)
-	assert.Regexp(t, `time="[^"]+" level=fatal msg="hello=world lorem=ipsum number=42" file="logger_test.go:[0-9]+"`, output)
+	assert.Regexp(t, `time="[^"]+" level=fatal msg="hello=world number=42" file="logger_inner_test.go:[0-9]+"`, output)
 	assert.Error(t, err, "exit status 1")
 }
